rework stride method names, don't expose stride check methods
