+++ src/lib.rs
@@ -78,9 +78,8 @@ use itertools::ZipSlices;
 pub use dimension::{
     Dimension,
     RemoveAxis,
-    can_index_slice,
-    dim_stride_overlap,
 };
+
 pub use dimension::NdIndex;
 pub use indexes::Indexes;
 pub use shape_error::ShapeError;
