+++ tests/dimension.rs
@@ -4,9 +4,6 @@ use ndarray::{
     Array,
     RemoveAxis,
     arr2,
-    can_index_slice,
-    dim_stride_overlap,
-    StrideError,
 };
 
 #[test]
@@ -39,25 +36,3 @@ fn dyn_dimension()
     assert_eq!(z.shape(), &dim[..]);
 }
 
-#[test]
-fn slice_indexing_uncommon_strides()
-{
-    let v: Vec<_> = (0..12).collect();
-    let dim = (2, 3, 2);
-    let strides = (1, 2, 6);
-    assert!(can_index_slice(&v, &dim, &strides).is_ok());
-
-    let strides = (2, 4, 12);
-    assert_eq!(can_index_slice(&v, &dim, &strides),
-               Err(StrideError::OutOfBoundsStride));
-}
-
-#[test]
-fn overlapping_strides_dim()
-{
-    let dim = (2, 3, 2);
-    let strides = (5, 2, 1);
-    assert!(dim_stride_overlap(&dim, &strides));
-    let strides = (6, 2, 1);
-    assert!(!dim_stride_overlap(&dim, &strides));
-}
